refactor(utxo-lib): consolidate `Transaction` type in test
Issue: BG-34381
