+++ modules/utxo-lib/test/integration_local_rpc/generate/types.ts
@@ -1,8 +1,21 @@
 /**
  * @prettier
  */
+
+import { Network } from '../../../src/networkTypes';
+import { Input } from '../../../src/bitgo/signature';
+
+type Output = {
+  value: number;
+  script: Buffer;
+};
+
 export type Transaction = {
+  network: Network;
+  ins: Input[];
+  outs: Output[];
   getId(): string;
+  hashForSignatureByNetwork(index: number, pubScript: Buffer, amount: number, hashType: number, isSegwit: boolean);
   toBuffer(): Buffer;
 };
